Quick changes to item duration code (#16)

+++ ValheimPlus/Configurations/Sections/ItemsConfiguration.cs
@@ -5,6 +5,6 @@
         public bool noTeleportPrevention { get; set; } = false;
         public float baseItemWeightReduction { get; set; } = 0;
         public float itemStackMultiplier { get; internal set; } = 1;
-        public float droppedItemOnGroundDurationInSeconds { get; internal set; } = 3600;
+        public int droppedItemOnGroundDurationInSeconds { get; internal set; } = 3600;
     }
 }
